pyproject.toml
[tool.poetry]
name = "pythonproject9"
version = "0.1.0"
description = ""
authors = ["Nastya Foxe <Kukusiki2204@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pillow = "^10.3.0"
django = "^5.0.6"
psycopg-binary = "^3.1.19"
djangorestframework = "^3.15.1"
psycopg = "^3.1.19"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"
black = "^24.4.2"
isort = "^5.13.2"
ipython = "^8.26.0"
drf-yasg = "^1.21.7"
setuptools = "^70.2.0"
django-cors-headers = "^4.4.0"
requests = "^2.32.3"
stripe = "^10.3.0"
forex-python = "^1.8"
python-dotenv = "^1.0.1"
pyopenssl = "^24.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
